+++ rust/src/exceptions.rs
@@ -5,6 +5,7 @@ use pyo3::prelude::*;
 create_exception!(_rustgrimp, ModuleNotPresent, PyException);
 create_exception!(_rustgrimp, NoSuchContainer, PyException);
 create_exception!(_rustgrimp, InvalidModuleExpression, PyException);
+create_exception!(_rustgrimp, CorruptCache, PyException);
 
 // We need to use here `pyclass(extends=PyException)` instead of `create_exception!`
 // since the exception contains custom data. See: